feat(create-webstone-app): add a system requirements task
