+++ packages/create-webstone-app/src/index.ts
@@ -1,6 +1,5 @@
-// FIXME: Does anyone know how to properly type the `pipe` function?
-/* eslint @typescript-eslint/no-explicit-any: "off" */
-import { tasks, displayWelcome, displayNextSteps } from "./helpers.js";
+import { displayWelcome, displayNextSteps } from "./helpers.js";
+import { tasks } from "./tasks/index.js";
 
 displayWelcome();
 const context = await tasks.run();
